@@ -16,47 +16,49 @@
     "lint": "ng lint"
   },
   "dependencies": {
-    "@angular/common": "^8.2.1",
-    "@angular/core": "^8.2.1",
-    "@angular/fire": "^5.2.1",
-    "@angular/forms": "^8.2.1",
-    "@angular/platform-browser": "^8.2.1",
-    "@angular/platform-browser-dynamic": "^8.2.1",
-    "@angular/pwa": "~0.802.1",
-    "@angular/router": "^8.2.1",
-    "@angular/service-worker": "^8.2.1",
-    "@google-cloud/tasks": "^1.8.0",
-    "@ionic-native/call-number": "^5.24.0",
-    "@ionic-native/camera": "^5.19.1",
-    "@ionic-native/chooser": "^5.21.6",
-    "@ionic-native/core": "^5.21.5",
-    "@ionic-native/crop": "^5.19.1",
-    "@ionic-native/document-viewer": "^5.22.0",
-    "@ionic-native/fcm": "^5.20.0",
-    "@ionic-native/file": "^5.19.1",
-    "@ionic-native/file-opener": "^5.19.1",
-    "@ionic-native/file-path": "^5.23.0",
-    "@ionic-native/file-transfer": "^5.22.0",
-    "@ionic-native/geolocation": "^5.19.1",
-    "@ionic-native/image-picker": "^5.19.1",
-    "@ionic-native/onesignal": "^5.19.1",
+    "@angular/common": "^9.1.4",
+    "@angular/core": "^9.1.4",
+    "@angular/fire": "6.0.0",
+    "@angular/forms": "^9.1.4",
+    "@angular/platform-browser": "^9.1.4",
+    "@angular/platform-browser-dynamic": "^9.1.4",
+    "@angular/router": "^9.1.4",
+    "@angular/service-worker": "^9.1.4",
+    "@google-cloud/tasks": "^1.9.0",
+    "@ionic-native/call-number": "^5.26.0",
+    "@ionic-native/camera": "^5.26.0",
+    "@ionic-native/chooser": "^5.26.0",
+    "@ionic-native/core": "^5.26.0",
+    "@ionic-native/crop": "^5.26.0",
+    "@ionic-native/document-viewer": "^5.26.0",
+    "@ionic-native/email-composer": "^5.26.0",
+    "@ionic-native/fcm": "^5.26.0",
+    "@ionic-native/file": "^5.26.0",
+    "@ionic-native/file-opener": "^5.26.0",
+    "@ionic-native/file-path": "^5.26.0",
+    "@ionic-native/file-transfer": "^5.26.0",
+    "@ionic-native/geolocation": "^5.26.0",
+    "@ionic-native/image-picker": "^5.26.0",
+    "@ionic-native/onesignal": "^5.26.0",
     "@ionic-native/social-sharing": "^5.26.0",
-    "@ionic-native/splash-screen": "^5.21.5",
-    "@ionic-native/status-bar": "^5.21.5",
-    "@ionic-native/toast": "^5.19.1",
-    "@ionic/angular": "5.0.0",
-    "@ngx-translate/core": "^11.0.1",
+    "@ionic-native/splash-screen": "^5.26.0",
+    "@ionic-native/status-bar": "^5.26.0",
+    "@ionic-native/toast": "^5.26.0",
+    "@ionic/angular": "^5.2.1",
+    "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
-    "angular-pipes": "^9.0.2",
+    "angular-pipes": "^10.0.0",
     "call-number": "^1.0.1",
     "cordova-android": "8.1.0",
     "cordova-ios": "^5.1.1",
     "cordova-plugin-androidx": "^1.0.2",
-    "cordova-plugin-androidx-adapter": "^1.1.0",
+    "cordova-plugin-androidx-adapter": "^1.1.1",
     "cordova-plugin-camera": "^4.1.0",
-    "cordova-plugin-chooser": "^1.2.6",
+    "cordova-plugin-chooser": "^1.3.1",
+    "cordova-plugin-crop": "^0.4.0",
     "cordova-plugin-document-viewer": "^0.9.13",
-    "cordova-plugin-fcm-with-dependecy-updated": "^4.1.1",
+    "cordova-plugin-email-composer": "^0.9.2",
+    "cordova-plugin-fcm-with-dependecy-updated": "^4.6.4",
     "cordova-plugin-file": "^6.0.2",
     "cordova-plugin-file-opener2": "^2.2.1",
     "cordova-plugin-file-transfer": "^1.7.1",
@@ -64,46 +66,46 @@
     "cordova-plugin-geolocation": "^4.0.2",
     "cordova-plugin-telerik-imagepicker": "git+https://github.com/Telerik-Verified-Plugins/ImagePicker.git",
     "cordova-plugin-x-socialsharing": "^6.0.0",
-    "core-js": "^2.5.7",
-    "dayjs": "1.8.0",
+    "core-js": "^2.6.11",
+    "dayjs": "^1.8.26",
     "es6-promise-plugin": "^4.2.2",
-    "firebase": "^7.2.3",
-    "firebase-admin": "^8.9.2",
-    "firebase-functions": "^3.3.0",
-    "google-libphonenumber": "^3.2.1",
-    "npm": "^6.13.7",
-    "rxjs": "6.5.2",
-    "tslib": "^1.10.0",
-    "videogular2": "6.4.0",
-    "zone.js": "~0.9.1"
+    "firebase": "^7.15.1",
+    "firebase-admin": "^8.12.1",
+    "firebase-functions": "^3.7.0",
+    "google-libphonenumber": "^3.2.10",
+    "npm": "^6.14.5",
+    "rxjs": "6.5.5",
+    "tslib": "^1.11.2",
+    "videogular2": "^7.0.2",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.802.1",
-    "@angular-devkit/build-angular": "^0.803.24",
-    "@angular-devkit/core": "^8.2.1",
-    "@angular-devkit/schematics": "^8.2.1",
-    "@angular/cli": "^8.2.1",
-    "@angular/compiler": "^8.2.1",
-    "@angular/compiler-cli": "^8.2.1",
-    "@angular/language-service": "^8.2.1",
+    "@angular-devkit/architect": "^0.901.8",
+    "@angular-devkit/build-angular": "^0.901.8",
+    "@angular-devkit/core": "^9.1.4",
+    "@angular-devkit/schematics": "^9.1.4",
+    "@angular/cli": "^9.1.4",
+    "@angular/compiler": "^9.1.4",
+    "@angular/compiler-cli": "^9.1.4",
+    "@angular/language-service": "~9.1.4",
     "@commitlint/cli": "^8.1.0",
     "@commitlint/config-angular": "^8.1.0",
-    "@ionic/angular-toolkit": "^2.0.0",
+    "@ionic/angular-toolkit": "^2.2.0",
     "@types/core-js": "^2.5.0",
     "@types/googlemaps": "^3.39.2",
-    "@types/node": "12.0.0",
-    "@webcomponents/webcomponentsjs": "^2.2.10",
-    "codelyzer": "^5.0.1",
-    "cordova-plugin-device": "^2.0.2",
+    "@types/node": "^12.11.1",
+    "@webcomponents/webcomponentsjs": "^2.4.3",
+    "codelyzer": "^5.2.2",
+    "cordova-plugin-device": "^2.0.3",
     "cordova-plugin-ionic-keyboard": "^2.2.0",
-    "cordova-plugin-ionic-webview": "^4.1.3",
-    "cordova-plugin-splashscreen": "^5.0.2",
-    "cordova-plugin-statusbar": "^2.4.2",
+    "cordova-plugin-ionic-webview": "^4.2.1",
+    "cordova-plugin-splashscreen": "^5.0.4",
+    "cordova-plugin-statusbar": "^2.4.3",
     "cordova-plugin-whitelist": "^1.3.4",
     "husky": "^1.3.1",
     "ts-node": "~8.1.0",
     "tslint": "~5.16.0",
-    "typescript": "~3.5.3"
+    "typescript": "^3.7.5"
   },
   "cordova": {
     "plugins": {
