@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
-
-import { NgMathPipesModule } from 'angular-pipes';
-import { FloorPipe } from 'angular-pipes';
+import { NgFloorPipeModule } from 'angular-pipes';
 
 import { TimeDifferencePipe } from './time-difference.pipe';
 import { TimeAgoPipe } from './time-ago.pipe';
@@ -12,17 +10,16 @@ import { TimeAgoPipe } from './time-ago.pipe';
   imports: [
     CommonModule,
     IonicModule,
-    NgMathPipesModule
+    NgFloorPipeModule
   ],
   declarations: [
     TimeDifferencePipe,
     TimeAgoPipe
   ],
   exports: [
-    FloorPipe,
+    NgFloorPipeModule,
     TimeDifferencePipe,
     TimeAgoPipe
-  ],
-  entryComponents: [],
+  ]
 })
 export class PipesModule {}
