@@ -1,5 +1,5 @@
 import { ShellModel } from '../../shell/data-store';
-import { PhotosArray } from '../../type';
+import { PhotosData } from '../../type';
 
 export class FirebaseListingItemModel extends ShellModel {
 
@@ -11,7 +11,7 @@ export class FirebaseListingItemModel extends ShellModel {
   object : string;
   description : string;
   coverPhotoData :string;
-  images : Array<PhotosArray> = [];
+  images : Array<PhotosData> = [];
   constructor() {
     super();
   }
