@@ -6,14 +6,14 @@ export const environment = {
   production: false,
   
   firebase:   {
-    apiKey: "AIzaSyBCvvDFnPEKABrn8EMgiTvOmVV5Gm7Etvw",
-    authDomain: "parkondo.firebaseapp.com",
-    databaseURL: "https://parkondo.firebaseio.com",
-    projectId: "parkondo",
-    storageBucket: "parkondo.appspot.com",
-    messagingSenderId: "373200357220",
-    appId: "1:373200357220:web:aaf7ddbcf978fff9b6dc71",
-    measurementId: "G-DQJTP89WRH" 
+    apiKey: 'AIzaSyBCvvDFnPEKABrn8EMgiTvOmVV5Gm7Etvw',
+    authDomain: 'parkondo.firebaseapp.com',
+    databaseURL: 'https://parkondo.firebaseio.com',
+    projectId: 'parkondo',
+    storageBucket: 'parkondo.appspot.com',
+    messagingSenderId: '373200357220',
+    appId: '1:373200357220:web:aaf7ddbcf978fff9b6dc71',
+    measurementId: 'G-DQJTP89WRH'
   }
   
  /* {
